test(activerecord): consolidate has-many-through stub into canonical file
Delete has-many-through.test.ts (orphan stub with 2 tests) and move
"loads through a join model" into the canonical
has-many-through-associations.test.ts. The stub's "get ids" test is
dropped because an equivalent (and more thorough) "get ids" test
already exists in the canonical file.
